Use tus for all uploads
Now that tusd supports uploading unknown length streams
Fixes #88

- Pull out sendTusRequest into sepearte module so it can be mocked
- Add argument chunkSize (defaults to 5MB only for unknown length streams, Infinity for file streams, same as tus-js-client)
- Emit deprecation warning when isResumable is set to false
