+++ README.md
@@ -149,6 +149,7 @@ You can provide the following keys inside the `options` object:
   - `uploadedBytes` - Number of bytes uploaded so far.
   - `totalBytes` - Total number of bytes to upload or `undefined` if unknown (Streams).
 - `onAssemblyProgress` - Once the Assembly has started processing this will be periodically called with the *Assembly Execution Status* (result of `getAssembly`) **only if `waitForCompletion` is `true`**.
+- `chunkSize` - (for uploads) a number indicating the maximum size of a TUS `PATCH` request body in bytes. Default to `Infinity` for file uploads and 5MB for streams of unknown length. See [tus-js-client](https://github.com/tus/tus-js-client/blob/master/docs/api.md#chunksize).
 
 Example code showing all options:
 ```js
